perpetual-inventory-app/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
databricks-sdk>=0.20.0
cachetools>=5.3.0
mlflow>=2.12.0
pydantic>=2.0.0
